requirements.txt
ultralytics
opencv-python
numpy
PyYAML
torch
torchvision
pandas
tqdm
pillow
pytest
matplotlib
seaborn
